+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# WIP
+
+- Install print and hashp to all frames in dynamic var bindings #16
+
 # 1.7.1 - Jan 15, 2026
 
 - Rename clj-kondo hooks namespace to avoid clashes #26 via @FelipeCortez